feat: partial precomputation support (#1)
Adds support for variable-time partial precomputation, which is useful
for multiscalar multiplication.

Currently, precomputation tables for use in multiscalar multiplication
must be generated and used in their entirety; this means it is not
possible to generate a superset of precomputation tables and then use
only a subset of them for computation without incurring the full
computational cost.

With this change, `vartime_mixed_multiscalar_mul` on a
`VartimeRistrettoPrecomputation` is allowed to use a static scalar
vector that is smaller than the static point vector. In this case, only
the static points corresponding to static scalars are used when
computing the multiscalar multiplication.

Also adds `Clone` support for precomputation table types, removes the
MSRV CI test, and adds build CI tests.